refactor: make more modular
